Fix local asset manager and duplicate loaders / savers being scanned
This fixes the local asset manager so it doesn't crash with a dependency
injection error when the asset manager starts.

It also fixes an issue where multiple loaders / savers / editors of the
same type could be registered if the location of the asset manager and
the current working directory were the same, and there were other
libraries that contained asset loaders / savers / editors to register.

Maniphest Tasks: T119