Disable triangulation in FBX reader; models should already be triangles before import
The conversion of quads to triangles sometimes causes issues with complex models (whose quads
can not be reasonably converted to triangles anyway).  When the triangulation occurs at an engine
level, it makes the appearance that the engine is not rendering the model correctly.  However, the
issues actually lie with the model's quads not being convertible to triangles without causing
rendering artifacts.  By disabling engine triangulation, we force these issues to be fixed before
the engine receives the model (e.g. in the modeling program with overlapping or incorrect
triangles can be corrected).