Implement IAutomaticAssetReload for reloading files on disk automatically
The IAutomaticAssetReload service can be used to automatically set up a
file system watcher on all of the source assets.  With this service, you
no longer need to explicitly dirty assets in your game when using
runtime proxies, instead the game will automatically reload assets when
they are changed on disk.