When Disconnect is called, block any reconnections from that endpoint
When a client and a server explicitly call Disconnect on each other's
endpoints, the packets that are still in transit can cause them to be
automatically reconnected (because Mx implicitly connects clients
when a packet is received).  Now when you call Disconnect, it'll prevent
any reconnections from occurring implicitly until Connect is called again
with that endpoints.