Remove Cloo as a dependency
It doesn't look like this library is being maintained any more, and given the range of compute APIs in addition to OpenCL (CUDA, DirectCompute), it doesn't seem to make much sense to keep maintaining this library.

I'd expect any work in future along the lines of compute shaders to be done upstream in MonoGame, since this area is a very hardware-abstraction section of the engine.