Disable multisampling by default on non-Windows platforms (#19)
After much investigation, the cause for render target corruption on macOS was the fact
that MonoGame does not update the MultiSampleCount property of the presentation
parameters once the OpenGL device is initialised.  Because this property is not initialised,
it means that requests to create render targets with the RenderTarget2D constructor get
passed a value of 32, which is invalid on almost every platform.  However, MonoGame (and
OpenGL) don't throw an error when passed a multisampling count too high for the render
target, so instead you just get a render target that is always corrupted and that you can't
do anything useful with.

Until MonoGame gets better support for multisampling on non-Windows platforms, we're
just going to leave multisampling off by default (developers can override PrepareDeviceSettings
in their game if they want to force it on).