Improvements to 3D deferred render pass to allow alpha blending with other passes
This allows a 3D deferred render pass to be rendered on top of a 2D render pass.  This can be used to
do cheap skyboxes by rendering the skybox as a texture and then rendering the 3D scene on top of
it (by placing the 3D deferred render pass later in the pipeline).