Update the audio handle so that it handles assets not being loaded on creation (#165)
This updates the audio handle so that you can create a handle on an audio asset that isn't yet loaded.  The audio asset will then create the audio instance on the next method call where the audio asset is also available.  In addition, it now checks to see if the audio asset is the same asset that the instance was created from (in the case of assets being reloaded), and automatically recreates the instance if that's the case.