Fix render pipeline tests due to SurfaceFormat change
The render pipeline tests use the SaveAsPng function to save the state of the render pipeline.  Due to a fix that was
previously made to support MSAA in the render pipeline, the render target being passed through to the tests
no longer has a SurfaceFormat of Color, and the SaveAsPng function doesn't support the new format.

This makes the tests blit the render target onto another render target whose SurfaceFormat is explicitly Color, thus
allowing SaveAsPng to work again.