Breaking changes: Remove host dependent classes from IGameContext, ICoreGame and IHostGame
Previously we provided GraphicsDeviceManager on ICoreGame and IGameContext.  However, this service is not
available if the underlying host is not derived from MonoGame's Game class, for example in the situation where
Protogame is being embedded inside another application (like the editor).

In addition, we also provided MonoGame's GameWindow (or platform specific variant) on ICoreGame and
IHostGame.  This is also not available if the game is not running from MonoGame's Game class.

This commit removes these properties in common API areas, so that developers won't accidentally use them
without knowing that they can sometimes be null.  They are still available with some casting required:

- GraphicsDeviceManager is available from `((HostGame)gameContext.Game.HostGame).GraphicsDeviceManager`
- GameWindow is available from `((Game)gameContext.Game.HostGame).Window`

Note that these casts will fail if the HostGame is not the correct implementation which makes these instances
available.

This commit also changes IRenderTargetBackBufferUtilities to accept IRenderContext instead of IGameContext.  This
API should never have accepted IGameContext since it deals strictly with rendering, but it was possible for
`GraphicsDevice` to be accessed via `gameContext.Graphics.GraphicsDevice`.  Since the GraphicsDeviceManager is
no longer available in IGameContext, this API had to be updated to use the GraphicsDevice provided on IRenderContext.
(Note that it's still possible to access the GraphicsDevice from IGameContext via the Game property, but this API needed
cleaning up anyway.)

Internally a bunch of platform-specific window resize detection and handling has been moved from CoreGame into
HostGame and internal references to GraphicsDeviceManager have been removed or replaced with the correct API
usage.  IGameWindow now has a Resize method which is used to resize the window on platforms that support it, though
developers should prefer the ResizeWindow method on IGameContext as that method also checks whether it's a valid
time to perform the resize.