Allow UI children to be scrollable-aware
This adds a new interface IScrollableAwareChild and an accompanying IScrollableAwareSkinRenderer interface for skin renderers.  When the scrollable container goes to render it's child, it will pass an additional rectangle indicating the area which will actually end up on screen.  This allows child controls to omit draw calls which are outside of the visible scrolled area.