Add interlaced depth API (#195)
This adds an IInterlacedBatchingDepthProvider which provides depth values for draw calls interlaced with a SpriteBatch in deferred mode.  If you use this interface to provide depth values to SpriteBatch (as I2DRenderUtilities does for you), you can then draw lines and triangles on the GPU directly while still having the SpriteBatch in deferred mode, which can save on draw calls.