Check to ensure the game is active before processing events (#196)
XNA/MonoGame captures all mouse and keyboard events, even when the window doesn't have focus.  This doesn't make a lot of sense (because it means you can click through other applications to the game window behind them), so this change ensures that the event engine hook checks to see if the game is active before firing any kind of event.