Skip render pipeline tests if no graphics device
Allow the render pipeline tests to be skipped only in the scenario where no graphics device is available.  Occasionally the build server loses the ability to use the graphics device (usually due to remoting into the service and killing the session attached to the physical screen).  When this occurs over the weekend, there's no way to restore session 0 so that the tests will pass again.  During this time, we want to allow builds to continue to pass.